feat(multi-schema): move baseFetch mock to test utils
